Correzione taskService frontend

+++ backend/machineMicroservice/src/main/java/com/mima/machine/service/TaskService.java
@@ -30,7 +30,7 @@ public interface TaskService {
      *
      * @return the list of entities
      */
-    List<TaskDTO> findAllByIdMachine(int idMachine);
+    List<TaskDTO> findAllByIdMachine(long idMachine);
 
 
     /**
